♻️ (webapp-svelte) Move some components to cdk
