+++ webapp-svelte/src/components/UserPublicInfo.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { IUser } from "@lib/user";
   import { classnames, dateFromObjectId } from "@/utils";
-  import { Card, CardBody, CardHeader } from "sveltestrap";
+  import { Card } from "@cdk";
 
   export let user: IUser
 
@@ -12,11 +12,8 @@
   $: joinDate = user && dateFromObjectId(user._id)
 </script>
 
-<Card class={classes}>
-  <CardHeader>About</CardHeader>
-  <CardBody>
-    ☯️ <a href="/page/karma" title="karma">{user.account.karma}</a> karma <br />
-    🎉 Joined us in {joinDate.toLocaleString("default", { month: "long" })}
-    {joinDate.toLocaleString("default", { year: "numeric" })}!
-  </CardBody>
+<Card class={classes} header="About">
+  ☯️ <a href="/page/karma" title="karma">{user.account.karma}</a> karma <br />
+  🎉 Joined us in {joinDate.toLocaleString("default", { month: "long" })}
+  {joinDate.toLocaleString("default", { year: "numeric" })}!
 </Card>
