+++ webapp-svelte/src/modules/cdk/index.ts
@@ -1,0 +1,5 @@
+export * from "sveltestrap";
+export { default as Card } from "./Card.svelte";
+export { default as Icon } from "./Icon.svelte";
+export { default as Loading } from "./Loading.svelte";
+export { default as Pagination } from "./Pagination.svelte";
