+++ webapp-svelte/src/utils/watch.ts
@@ -1,27 +1,10 @@
 /**
- * Allows to call an expression each time deps are modified, and once initially
- *
- * @param callback
- * @param deps
- *
- * @example
- *
- * ```
- * const immediate = () => console.log("executed immediately and on changes")
- *
- * $ : watch(immediate, [a, b, c])
- * ```
- */
-export function watch(callback: () => unknown, _deps: any[]) {
-  callback();
-}
-
-/**
- * Create a watcher. Compared to `watch`, it can have additional options
+ * Create a watcher. Compared to `watch`, it can have additional options and
+ * by default only executes after changes, not the first time
  *
  * @param callback The function to executre when the deps change
  * @param options `options.immediate` is whether or not the function is executed
- * immediately or only on changes
+ * immediately or only on changes. Defaults to false.
  * @returns The watcher. Pass it the deps.
  *
  * @example
@@ -30,10 +13,15 @@ export function watch(callback: () => unknown, _deps: any[]) {
  * const watcher = createWatcher(() => console.log("executed only on changes on a/b/c"), {immediate: false})
  *
  * $ : watcher(a, b, c)
+ * // or
+ * $ : watcher(), [a,b,c]
  * ```
  */
-export function createWatcher(callback: () => unknown, options: { immediate: boolean }) {
-  let skip = !options.immediate;
+export function createWatcher(
+  callback: () => unknown,
+  { immediate = false }: { immediate: boolean } = { immediate: false }
+) {
+  let skip = !immediate;
 
   return (..._deps: any[]) => {
     if (skip) {
